Rename Faberbabel object to LocalizableManager
