+++ Faberbabel/Classes/LocalizableManager+Translation.swift
@@ -1,5 +1,5 @@
 //
-//  Faberbabel+Translation.swift
+//  LocalizableManager+Translation.swift
 //  Faberbabel
 //
 //  Created by Pierre Felgines on 08/09/2020.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Faberbabel {
+extension LocalizableManager {
 
     func translation(forKey key: String,
                      lang: String) -> String {
